fix: extract localhost URL click handler, fix TaskExecDetail missing handler
TaskExecDetail.vue lacked localhost URL click handling — clicking localhost
links navigated directly instead of going through SSH tunnel + WebView,
causing redirect to login page in Android App mode.

Extract shared localhost URL click logic into useLocalhostUrlClickHandler()
composable in useLocalhostAnnotation.ts, replacing 3 copies of identical
code in ChatMessageList, ToolDetailOverlay, and TaskExecDetail.
